refactor!: separate "save" into its own directory
