+++ raillabel/__init__.py
@@ -8,7 +8,7 @@ from .exceptions import *
 from .filter.filter import filter
 from .format.scene import Scene
 from .load_.load import load
-from .save import save
+from .save.save import save
 from .validate.validate import validate
 
 try:
