+++ raillabel/save/save.py
@@ -4,9 +4,9 @@
 import json
 from pathlib import Path
 
-from . import exceptions
-from .format import Scene
-from .validate.validate import validate as validate_func
+from .. import exceptions
+from ..format import Scene
+from ..validate.validate import validate as validate_func
 
 
 def save(scene: Scene, path: str, prettify_json: bool = False, validate: bool = False):
